QUICK-692 run pg schema migrations before couchbase migrations (#39)
What
Changes the order of the api startup flow so that postgres schema migrations run before couchbase migrations.

Why
The runbooks v2 feature is built to rely on the couchbase migration runner to seed data into postgres. While this isn't ideal, we are going to support this for now because we don't have another alternative in place for the Oaks team to use to do this data seeding at the moment. So in order for the couchbase migration runner to be able to seed data into the runbooks postgres tables, the tables will need to exist. Therefore, we now need the pg schema migrations to run before the couchbase migrations.