PG-77 Added a volume for localesOverride (to support white-labeling) (#78)
Added `localesOverride` volume to the `plextracapi` container.

This volume allows the "white-labeling" feature to persist changes between container restarts.  

This is <b><i>part of the change</i></b> necessary to eliminate the need for the CS team to manually update new PlexTrac instances to enable the "White Labeling" feature.

Prior to this change, the CS team needed to manually apply these changes to every new PlexTrac environment:

https://plextrac.atlassian.net/wiki/spaces/CE/pages/2330820609/Whitelabeling+-+Cannot+Wrap+An+Error